Bug fixing: Bug fixing and cleanup
Handle the case of a missing database better. Add also the option to
explicitly initialize the database.
Remove the hardcoded path for the DB and use proper home directory
expansion to find it.
Add a cleaner way to print debug messages.
Fix reporting of time. Reporting was printing seconds instead of
minutes.